requirements.txt
tensorflow
numpy 
keras.preprocessing 
